Add event unwrapper: s3-in-sns
A common pattern in Lambda functions is to be subscribed to an SNS topic
that is sending out S3 notification events. Getting down to the meat of
the S3 event(s) that triggered the Lambda function requires quite a bit
of parsing through the event structure(s), so this unwrapper makes that
into a reusable function.